compact.c: options du runtime -> unsigned long            arrondi dans le calcul de pourcentage            changement de max_overhead config.h: changement de max_overhead freelist.c: meilleure defragmentation gc_ctrl.c: meilleure defragmentation; options du runtime -> unsigned long;            arrondi dans les calculs de heap_increment gc_ctrl.c: options du runtime -> unsigned long; major_gc.c: changement de space_overhead stacks.c: message pour afficher max_stack_size startup.c: options du runtime -> unsigned long
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@1568 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ byterun/gc_ctrl.h
@@ -25,7 +25,8 @@ extern long
      stat_heap_size,
      stat_compactions;
 
-void init_gc P((unsigned long, unsigned long, unsigned long, int, int, int));
+void init_gc P((unsigned long, unsigned long, unsigned long,
+		unsigned long, unsigned long, unsigned long));
 
 
 #endif /* _gc_ctrl_ */
